requirements.txt
python-telegram-bot==22.3
google-generativeai==0.8.5
python-dotenv==1.1.1